🔍 기타 사항 리펙토링
assembly.js : 실질적인 HTML의 수정 작업을 담당하기 위해 리펙토링을 진행했다.
clickEventHandler.js : 이후 진행될 클릭 이벤트의 증가를 위해 기능 일부를 파일로 새로 생성했다.
onNavClickHandler.js : 네비게이션의 클릭 이벤트를 담당하기 위해 생성했다.
parts.js : HTML의 생성 작업을 담당하기 위해 직접적으로 이 파일에서만 데이터를 받게 수정했다.
